+++ Model/Service/Sync/Upsert/Product/SyncService.php
@@ -34,13 +34,12 @@
  *
  */
 
-namespace Nosto\Tagging\Model\Service\Sync\Upsert;
+namespace Nosto\Tagging\Model\Service\Sync\Upsert\Product;
 
 use Exception;
 use Magento\Catalog\Model\Product;
 use Magento\Store\Model\Store;
 use Nosto\Exception\MemoryOutOfBoundsException;
-use Nosto\Tagging\Model\ResourceModel\Magento\Category\Collection as CategoryCollection;
 use Nosto\NostoException;
 use Nosto\Operation\UpsertProduct;
 use Nosto\Request\Http\Exception\AbstractHttpException;
@@ -51,12 +50,9 @@ use Nosto\Tagging\Logger\Logger as NostoLogger;
 use Nosto\Tagging\Model\ResourceModel\Magento\Product\Collection as ProductCollection;
 use Nosto\Tagging\Model\Service\AbstractService;
 use Nosto\Tagging\Model\Service\Cache\CacheService;
-use Nosto\Tagging\Model\Service\Product\Category\CategoryServiceInterface as NostoCategoryService;
 use Nosto\Tagging\Model\Service\Product\ProductServiceInterface;
 use Nosto\Tagging\Util\PagingIterator;
 use Nosto\Tagging\Model\Product\Repository as ProductRepository;
-use Nosto\Tagging\Model\Category\Repository as CategoryRepository;
-use Nosto\Tagging\Model\Operation\CategoryUpdateOnNosto;
 
 class SyncService extends AbstractService
 {
@@ -87,11 +83,6 @@ class SyncService extends AbstractService
     /** @var ProductRepository */
     private ProductRepository $productRepository;
 
-    private CategoryRepository $categoryRepository;
-
-    /** @var NostoCategoryService */
-    private NostoCategoryService $nostoCategoryService;
-
     /**
      * Sync constructor.
      * @param NostoHelperAccount $nostoHelperAccount
@@ -101,8 +92,6 @@ class SyncService extends AbstractService
      * @param ProductServiceInterface $productService
      * @param CacheService $cacheService
      * @param ProductRepository $productRepository
-     * @param CategoryRepository $categoryRepository
-     * @param CategoryServiceInterface $nostoCategoryService
      * @param $apiBatchSize
      * @param $apiTimeout
      */
@@ -114,8 +103,6 @@ class SyncService extends AbstractService
         ProductServiceInterface $productService,
         CacheService $cacheService,
         ProductRepository $productRepository,
-        CategoryRepository $categoryRepository,
-        NostoCategoryService $nostoCategoryService,
         $apiBatchSize,
         $apiTimeout
     ) {
@@ -126,8 +113,6 @@ class SyncService extends AbstractService
         $this->nostoDataHelper = $nostoDataHelper;
         $this->cacheService = $cacheService;
         $this->productRepository = $productRepository;
-        $this->categoryRepository = $categoryRepository;
-        $this->nostoCategoryService = $nostoCategoryService;
         $this->apiBatchSize = $apiBatchSize;
         $this->apiTimeout = $apiTimeout;
     }
@@ -140,7 +125,7 @@ class SyncService extends AbstractService
      * @throws AbstractHttpException
      * @throws Exception
      */
-    public function syncProducts(ProductCollection $collection, Store $store)
+    public function sync(ProductCollection $collection, Store $store)
     {
         if (!$this->nostoDataHelper->isProductUpdatesEnabled($store)) {
             $this->logDebugWithStore(
@@ -152,7 +137,6 @@ class SyncService extends AbstractService
         $account = $this->nostoHelperAccount->findAccount($store);
         $this->startBenchmark(self::BENCHMARK_SYNC_NAME, self::BENCHMARK_SYNC_BREAKPOINT);
 
-        $index = 0;
         $collection->setPageSize($this->apiBatchSize);
         $iterator = new PagingIterator($collection);
 
@@ -168,7 +152,6 @@ class SyncService extends AbstractService
                     ($iterator->getCurrentPageNumber() - 1) * $this->apiBatchSize
                 )
             );
-            $index ++;
             /** @var Product $product */
             foreach ($products->getItems() as $product) {
                 $productIdsInBatch[] = $product->getId();
@@ -194,77 +177,5 @@ class SyncService extends AbstractService
             $op->upsert();
         }
         $this->logBenchmarkSummary(self::BENCHMARK_SYNC_NAME, $store, $this);
-    }
-
-    /**
-     * @throws NostoException
-     * @throws MemoryOutOfBoundsException
-     */
-    public function syncCategories(CategoryCollection $collection, Store $store)
-    {
-        if (!$this->nostoDataHelper->isProductUpdatesEnabled($store)) {
-            $this->logDebugWithStore(
-                'Nosto Category sync is disabled',
-                $store
-            );
-            return;
-        }
-        $categoryIdsInBatch = [];
-        $account = $this->nostoHelperAccount->findAccount($store);
-        $this->startBenchmark('nosto_category_upsert', self::BENCHMARK_SYNC_BREAKPOINT);
-        $collection->setPageSize($this->apiBatchSize);
-        $iterator = new PagingIterator($collection);
-
-        /** @var CategoryCollection $page */
-        foreach ($iterator as $page) {
-            $this->checkMemoryConsumption('category sync');
-            foreach ($page->getData() as $category) {
-                $categoryDb = $this->categoryRepository->getByIds($category['entity_id']);
-                $categoryIdsInBatch[] = $category['entity_id'];
-                $id = $category['entity_id'];
-                $name = explode("/", $categoryDb->getName());
-                $name = end($name);
-                $parentId = $categoryDb->getParentId();
-                $urlPath = $this->getUrlPath($categoryDb, $store);
-                $fullName = $categoryDb->getName();
-                $available = $categoryDb->getIsActive() ?? false;
-                try {
-                    (new CategoryUpdateOnNosto($account, $store->getCurrentUrl(), $id, $name, $parentId, $urlPath, $fullName, $available))->execute();
-                } catch (Exception $e) {
-                    $this->logger->logDebugWithStore(sprintf(
-                        'Failed to update Nosto category %s with %s id',
-                        $name,
-                        $id
-                    ));
-                }
-            }
-
-            $this->logDebugWithStore(
-                sprintf(
-                    'Nosto category sync with ids - %s',
-                    implode(',', $categoryIdsInBatch)
-                ),
-                $store
-            );
-        }
-    }
-
-    private function getUrlPath($category, Store $store)
-    {
-        $nostoCategory = '';
-        try {
-            $data = [];
-            $path = $category->getPath();
-            $data[] = $category->getName();
-            $nostoCategory = count($data) ? '/' . implode('/', $data) : '';
-        } catch (Exception $e) {
-            $this->logDebugWithStore($e, $store);
-        }
-
-        if (empty($nostoCategory)) {
-            $nostoCategory = null;
-        }
-
-        return $nostoCategory ? trim($nostoCategory) : $nostoCategory;
     }
 }
